+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Implement the registration GUI.
 * Remove the event-based client API for registration, instead having the GUI use the async method from
   `RegistrationClient` directly.
+* Implement the login GUI.
 
 #### 23 February 2024